+++ src/helpers/GoogleMaps.php
@@ -13,6 +13,7 @@ namespace doublesecretagency\googlemaps\helpers;
 
 use Craft;
 use craft\base\Element;
+use craft\elements\ElementCollection;
 use doublesecretagency\googlemaps\GoogleMapsPlugin;
 use doublesecretagency\googlemaps\models\DynamicMap;
 use doublesecretagency\googlemaps\models\Location;
@@ -97,11 +98,11 @@ class GoogleMaps
     /**
      * Create a new Dynamic Map object.
      *
-     * @param array|Element|Location $locations
+     * @param array|ElementCollection|Element|Location $locations
      * @param array $options
      * @return DynamicMap
      */
-    public static function map(array|Element|Location $locations = [], array $options = []): DynamicMap
+    public static function map(array|ElementCollection|Element|Location $locations = [], array $options = []): DynamicMap
     {
         // Create a new map object
         $map = new DynamicMap($locations, $options);
@@ -142,11 +143,11 @@ class GoogleMaps
     /**
      * Create a new Static Map object.
      *
-     * @param array|Element|Location $locations
+     * @param array|ElementCollection|Element|Location $locations
      * @param array $options
      * @return StaticMap
      */
-    public static function img(array|Element|Location $locations = [], array $options = []): StaticMap
+    public static function img(array|ElementCollection|Element|Location $locations = [], array $options = []): StaticMap
     {
         return new StaticMap($locations, $options);
     }